group-manager-python/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure SDK (for Step 2)
azure-identity==1.15.0
msgraph-sdk==1.0.0

# Validation
email-validator==2.1.0

# Logging
python-json-logger==2.0.7

# Python utilities
python-dotenv==1.0.0
